a development branch was accidentally merged in to master, this is why someone else should review PRs

+++ src/sdlGraphicsInterface.hpp
@@ -34,7 +34,7 @@ private:
     int texturePitch;
 
     SDL_FPoint topLeft;
-    double cellSize;
+    float cellSize;
 
     SDL_Event event;
